extraced _OVERLAPPED
